# requests/OpenAPI/requests.yaml
openapi: 3.0.3
info:
  title: Bird — Twitter Clone API
  description: |
    Two-service microservice architecture:
    - **UMS** (User Management Service) — port `9000`
    - **Twitter** (Messaging & Subscriptions Service) — port `9001`

    The Twitter service calls UMS internally to validate user roles before
    every write operation, so UMS must be running first.
  version: "2.0"
  contact:
    name: Eugene Zimin
  license:
    name: BSD 2-Clause
    url: https://opensource.org/licenses/BSD-2-Clause

# Two logical servers — one per microservice
servers:
  - url: http://localhost:9000
    description: UMS — User Management Service
  - url: http://localhost:9001
    description: Twitter — Messaging & Subscriptions Service
...
paths:
...
  /users/user/{userId}:
    servers:
      - url: http://localhost:9000
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags: [UMS / Users]
      summary: Get a user by ID
      operationId: getUserById
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeUser"
              example:
                code: "200"
                message: User retrieved successfully
                data:
                  id: 6e27ea06-a716-4c89-af88-813749a8bd48
                  name: Donald Trump
                  email: donald@trump.us
                  createdAt: "2020-07-31T08:00:00"
                  roles:
                    - roleId: eb932dbb-7005-422f-a649-7190af39e984
                      role: PRODUCER
                      description: Message content producer
                  lastSession:
                    sessionId: 56de00c1-a7dd-4f66-9728-307c5230455e
                    loggedInAt: "2020-11-12T13:55:00"
                    loggedOutAt: "2020-11-12T15:00:00"
        "200x":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeEmpty"
              example:
                code: "404"
                message: User not found
                data: {}
    put:
      tags: [UMS / Users]
      summary: Update a user (full replacement)
      description: >
        Replaces all mutable fields (name, email, passwordHash) and re-syncs
        the role list entirely. Send every field you want to keep — absent
        fields are not preserved. Use PATCH for partial updates.
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
            example:
              name: Donald J. Trump
              email: donald@trump.us
              passwordHash: newpassword
              roles:
                - role: SUBSCRIBER
                - role: PRODUCER
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeBoolean"
              examples:
                success:
                  value:
                    code: "200"
                    message: User updated
                    data: true
                notFound:
                  value:
                    code: "404"
                    message: User not found or not updated
                    data: false
    patch:
      tags: [UMS / Users]
      summary: Patch a user (partial update)
      description: >
        Applies only the fields present in the request body. Absent fields are
        left unchanged. Roles are **not** affected by PATCH — use PUT to
        replace the role list.  At least one of `name`, `email`, or
        `passwordHash` must be provided; an empty body returns 400.
        On success the response contains the refreshed user object.
      operationId: patchUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
            examples:
              nameOnly:
                summary: Change display name only
                value:
                  name: Donald J. Trump
              emailAndPassword:
                summary: Rotate email and password together
                value:
                  email: djt@trump.us
                  passwordHash: supersecret123
      responses:
        "200":
          description: Patch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeUser"
              examples:
                success:
                  value:
                    code: "200"
                    message: User 6e27ea06-a716-4c89-af88-813749a8bd48 patched successfully
                    data:
                      id: 6e27ea06-a716-4c89-af88-813749a8bd48
                      name: Donald J. Trump
                      email: donald@trump.us
                      createdAt: "2020-07-31T08:00:00"
                      roles:
                        - roleId: eb932dbb-7005-422f-a649-7190af39e984
                          role: PRODUCER
                          description: Message content producer
                notFound:
                  value:
                    code: "404"
                    message: User 6e27ea06-a716-4c89-af88-813749a8bd48 not found
                    data: false
                emptyBody:
                  value:
                    code: "400"
                    message: "Request body must contain at least one of: name, email, passwordHash"
                    data: false
    delete:
      tags: [UMS / Users]
      summary: Delete a user
      operationId: deleteUser
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeBoolean"
              examples:
                success:
                  value:
                    code: "200"
                    message: User deleted
                    data: true
                notFound:
                  value:
                    code: "404"
                    message: User not found or already deleted
                    data: false
  /users/user/{userId}/role/{roleId}:
    servers:
      - url: http://localhost:9000
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/roleId"
    post:
      tags: [UMS / Roles]
      summary: Assign a role to a user
      description: >
        Assigns an existing role to an existing user. The operation is
        idempotent — calling it when the role is already assigned returns
        200 without error. Both the user and the role must exist; otherwise
        404 is returned inside the envelope.
      operationId: assignRole
      responses:
        "200":
          description: Assignment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeBoolean"
              examples:
                success:
                  value:
                    code: "200"
                    message: Role assigned to user 6e27ea06-a716-4c89-af88-813749a8bd48
                    data: true
                userNotFound:
                  value:
                    code: "404"
                    message: User 6e27ea06-a716-4c89-af88-813749a8bd48 not found
                    data: false
                roleNotFound:
                  value:
                    code: "404"
                    message: Role b479b357-7e25-47fa-8dba-bfdaeecc6c2c not found
                    data: false

  # =========================================================================
  # UMS — Roles
  # =========================================================================
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 6e27ea06-a716-4c89-af88-813749a8bd48
    roleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: b479b357-7e25-47fa-8dba-bfdaeecc6c2c
...
  schemas:
...
    Role:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
        role:
          type: string
          enum: [ADMIN, PRODUCER, SUBSCRIBER]
        description:
          type: string
    LastSession:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        loggedInAt:
          type: string
          format: date-time
        loggedOutAt:
          type: string
          format: date-time
          nullable: true
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        lastSession:
          $ref: "#/components/schemas/LastSession"
    UserCreate:
      type: object
      required: [name, email, passwordHash]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        passwordHash:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    UserPatch:
      type: object
      description: >
        Partial user update. At least one field must be present.
        Absent fields are left unchanged. Roles are not affected.
      properties:
        name:
          type: string
          example: Donald J. Trump
        email:
          type: string
          format: email
          example: djt@trump.us
        passwordHash:
          type: string
          example: supersecret123
...
    EnvelopeUser:
      type: object
      properties:
        code:    { type: string }
        message: { type: string }
        data:
          $ref: "#/components/schemas/User"
...
    EnvelopeBoolean:
      type: object
      properties:
        code:    { type: string }
        message: { type: string }
        data:
          type: boolean
...
    EnvelopeEmpty:
      type: object
      properties:
        code:    { type: string }
        message: { type: string }
        data:
          type: object